# vectorize string-escape with NEON